skip more tests when using unsupported dns resolvers
